Move tests folder top level to avoid having them in published npm files
