+++ tests.node.js
@@ -1,3 +1,3 @@
 global.__DEV__ = true
 
-require('./modules/__tests__/serverRendering-test')
+require('./tests/serverRendering-test')
