+++ tests/IndexLink-test.js
