+++ tests/IndexRedirect-test.js
