+++ tests/applyRouterMiddleware-test.js
