+++ tests/formatPattern-test.js
