+++ tests/getParams-test.js
