+++ tests/isActive-test.js
