+++ tests/matchPattern-test.js
