+++ tests/shouldWarn.js
